requirements.txt
customtkinter==5.2.2
Pillow==10.1.0
CTkMessagebox==2.5
emoji==2.10.1
config==0.5.1